+++ zingolib/src/data.rs
@@ -18,7 +18,7 @@ pub mod receivers {
     use zcash_client_backend::zip321::Payment;
     use zcash_client_backend::zip321::TransactionRequest;
     use zcash_client_backend::zip321::Zip321Error;
-    use zcash_primitives::memo::MemoBytes;
+    use zcash_protocol::memo::MemoBytes;
     use zcash_protocol::value::Zatoshis;
 
     /// A list of Receivers