+++ zingolib/src/wallet/zcb_traits.rs
@@ -17,7 +17,6 @@ use zcash_client_backend::{
 use zcash_keys::{address::UnifiedAddress, keys::UnifiedFullViewingKey};
 use zcash_primitives::{
     block::BlockHash,
-    legacy::{TransparentAddress, keys::TransparentKeyScope},
     memo::Memo,
     transaction::{Transaction, TxId},
 };
@@ -25,7 +24,9 @@ use zcash_protocol::{
     PoolType, ShieldedProtocol,
     consensus::{self, BlockHeight, Parameters},
 };
+use zcash_transparent::address::TransparentAddress;
 use zcash_transparent::bundle::{OutPoint, TxOut};
+use zcash_transparent::keys::TransparentKeyScope;
 
 use super::{LightWallet, error::WalletError, output::OutputRef};
 use crate::wallet::output::RemainingNeeded;
@@ -434,14 +435,18 @@ impl WalletWrite for LightWallet {
         for sent_transaction in transactions {
             // this is a workaround as Transaction does not implement Clone
             let mut transaction_bytes = vec![];
-            sent_transaction.tx().write(&mut transaction_bytes)?;
+            sent_transaction
+                .tx()
+                .write(&mut transaction_bytes)
+                .map_err(WalletError::TransactionWrite)?;
             let transaction = Transaction::read(
                 transaction_bytes.as_slice(),
                 consensus::BranchId::for_height(
                     &self.network,
                     sent_transaction.target_height().into(),
                 ),
-            )?;
+            )
+            .map_err(WalletError::TransactionRead)?;
 
             match pepper_sync::scan_pending_transaction(
                 &network,